requirements.txt
argcomplete==1.12.2
asgiref==3.3.1
boto3==1.16.55
botocore==1.19.55
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
Django==3.1.5
django-appconf==1.0.4
django-cors-headers==3.4.0
django-environ==0.4.5
django-extensions==3.0.4
django-filter==2.3.0
django-simple-history==2.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
jmespath==0.10.0
kappa==0.6.0
mysqlclient==2.0.1
pip-tools==5.5.0
placebo==0.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.56.0
troposphere==2.6.3
urllib3==1.26.2
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zappa-django-utils==0.4.1
